TextToSQL-Service/requirements.txt
uvicorn
fastapi~=0.115.8
pydantic~=2.10.6
torch~=2.6.0
transformers~=4.48.3
datasets~=3.3.0
aioredis
redis
sentencepiece
openai~=1.63.0

pip
attrs
distro
filelock
requests~=2.32.3
pytest-bdd~=8.1.0
huggingface-hub~=0.28.1
python-dotenv~=1.0.1